Rename: retrieve subgraph directory
